requirements.txt
# pip3 install -U -r requirements.txt
numpy
opencv-python
torch == 1.2.0
matplotlib
pycocotools
tqdm
tb-nightly
future
Pillow

# Equivalent conda commands ----------------------------------------------------
# conda update -n base -c defaults conda
# conda install -yc anaconda future numpy opencv matplotlib tqdm pillow
# conda install -yc conda-forge scikit-image tensorboard pycocotools
# conda install -yc spyder-ide spyder-line-profiler
# conda install -yc pytorch pytorch torchvision
